commit before changes to return values of filters
Since we are planning to cache some responses from the backend, we need to change the return values of before filters to *res and method name. When after filters are impemented, before filters will ask them if they have cached the answers for the req, if so they immediately set the response.